% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliqutils.R
\name{areCliques}
\alias{areCliques}
\title{Check if membership vector encodes proper cliques for a given adjacency relationship encoding matrix}
\usage{
areCliques(WorA, cl_mem)
}
\arguments{
\item{WorA}{A numeric matrix containing adjacency relationship information between nodes. \code{WorA[i,j]>0} means that nodes \code{i,j} are connected.}

\item{cl_mem}{An integer membership vector for wihch \code{cl_mem[i]} gives a label of the clique to which node \code{i} belongs. A value of \code{0} can possibly encode omitted nodes, for which the check is not performed.}
}
\value{
\code{TRUE} if all groups of nodes described by \code{cl_mem} are valid cliques (including pairs and singletons), \code{FALSE} otherwise.
}
\description{
Assuming a input matrix encodes ajdacency relation or weights where \code{0} means no connection, function checks whether the membership vector provided properly partitions nodes into a set of valid cliques.
}
\examples{
W<- matrix(runif(40*40), ncol=40) +1
W= W \%*\% t(W)
membership= c( rep(1,3), rep(2, ncol(W) - 3) )
areCliques(W, membership)
W[3,1]<-W[1,3]<-0
areCliques(W, membership)
membership[1:3]<-0
areCliques(W, membership)
}
