% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliqutils.R
\name{fastTable}
\alias{fastTable}
\title{A minimal, efficient table function for one dimensional vectors}
\usage{
fastTable(values)
}
\arguments{
\item{values}{A 1 dimensional atomic vector}
}
\value{
A \code{list} wirth the following elements: 
\itemize{
\item \code{value} a vector of unique values in \code{values}.
\item \code{count} a vector of counts of each element of \code{value} component, i.e. \code{fastTable(x)$count[i]} is frequency of \code{fastTable(x)$value[i]} in \code{x}.
	   }
}
\description{
Emulates \code{table} in an efficient manner for a narrow case of one dimensional atomic vectors. Note that no correctness checks are done.
}
\examples{
fastTable( sample(c(1,2,3,4),100,replace=TRUE,prob=c(0.2,0.1,0.5,0.2)))
}
