% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliqutils.R
\name{isClique}
\alias{isClique}
\title{Check if an adjacency or weight matrix encodes a clique}
\usage{
isClique(WorA, subgr = NULL)
}
\arguments{
\item{WorA}{A numeric matrix containing adjacency relationship information between nodes. \code{WorA[i,j]>0} means that nodes \code{i,j} are connected.}

\item{subgr}{An indexing vector specifying a subraph of the nodes for which to check if they are forming a clique. If \code{NULL}, check is done using all nodes in the input graph.}
}
\value{
\code{TRUE} if the graph in question is a clique, \code{FALSE} otherwise.
}
\description{
Assuming a input matrix encodes ajdacency relation or weights where \code{0} means no connection, function checks whether the nodes described by the matrix are forming a valid clique. Can also be used by specifying a subgraph of the input graph.
}
\examples{
W<- matrix(runif(40*40), ncol=40) + 1
W= W \%*\% t(W)
isClique(W)
W[3,1]<-W[1,3]<-0
isClique(W)
isClique(W, 4:ncol(W))
}
