% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliqutils.R
\name{max2minOrder}
\alias{max2minOrder}
\title{Order values in the decreasing manner}
\usage{
max2minOrder(x)
}
\arguments{
\item{x}{A vector to be ordered}
}
\value{
A permutation of indexes of \code{x} that sorts the vector in the decreasing manner.
}
\description{
Order values in the decreasing manner
}
\examples{
y<- runif(10)*10
ord_y<- max2minOrder(y)
print(y)
print(ord_y) 
print(y[ord_y])
}
