% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpplique_gw_sumPC}
\alias{rcpplique_gw_sumPC}
\title{Rcpp Implementation of Basic Greedy Clique Partitioner}
\usage{
rcpplique_gw_sumPC(W_r)
}
\arguments{
\item{W_r}{A numeric matrix, assumed to be symmetric with zeroes on its diagonal.}
}
\value{
An integer vector encoding the clique membership of each node. Zeroes encode singletons.
}
\description{
This function assumes the matrix is symmetric and its diagonal is zeroed out
and performs no checks for correctness of the input.
Use \code{greedyCliquePartitioner()} for a robust version.
}
