% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliqutils.R
\name{tidyUpLabels}
\alias{tidyUpLabels}
\title{Relabel group labels to consecutive integers}
\usage{
tidyUpLabels(labelV)
}
\arguments{
\item{labelV}{An integer membership vector for which \code{cl_mem[i]} gives a label of the group to which node \code{i} belongs. A value of \code{0} by assumption encodes nodes that will be not taken into the account by the relabeling operation.}
}
\value{
An altered integer membership vector in which each label corresponding to the group of size \code{ < mS } is changed to \code{0}.
}
\description{
Based on input integer label vector, normalize labels such that if there are \code{M} unique labels in total, each of them gets mapped to one of the integers in \code{1:M}.
}
\examples{
mem<- c(0,0,0,1,0,1,67,3,0,1,67,67,5,5,4,4,0,0)
tidyUpLabels(mem)
}
