% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliqutils.R
\name{uniqueSingletonLabels}
\alias{uniqueSingletonLabels}
\title{Relabel singleton nodes from \code{0} to unique values}
\usage{
uniqueSingletonLabels(cl_mem)
}
\arguments{
\item{cl_mem}{An integer membership vector for which \code{cl_mem[i]} gives a label of the group to which node \code{i} belongs. A value of \code{0} by assumption encodes singleton nodes.}
}
\value{
An altered integer membership vector in which each singleton (formerly encoded by \code{0}) gets a new, unique numerical label.
}
\description{
Function is interpreted in the contex where input vector describes partition of nodes into groups. In general it performs a simple substitution of \code{0} values by unique integers.
}
\examples{
mem<- c(0,0,0,1,0,1,2,3,0,1,2,2,2,1,4,4,0,0)
uniqueSingletonLabels(mem)
}
