% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliqutils.R
\name{zeroOutSmall}
\alias{zeroOutSmall}
\title{Relabel small groups to \code{0}}
\usage{
zeroOutSmall(labelV, mS)
}
\arguments{
\item{labelV}{An integer membership vector for which \code{cl_mem[i]} gives a label of the group to which node \code{i} belongs. A value of \code{0} by assumption encodes nodes that will be not taken into the account by the relabeling operation.}

\item{mS}{Integer scalar giving minimum group size that is not relabeled to \code{0}.}
}
\value{
An altered integer membership vector in which each label corresponding to the group of size \code{ < mS } is changed to \code{0}.
}
\description{
Based on input label vector, zero out labels corresponding to groups smaller than given minimal group size.
}
\examples{
mem<- c(0,0,0,1,0,1,2,3,0,1,2,2,2,1,4,4,0,0)
zeroOutSmall(mem,3)
}
